requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
pillow>=8.0.0
insightface>=0.6.2
onnxruntime>=1.8.0
opencv-python>=4.5.3
matplotlib>=3.4.0
wandb>=0.12.0
tqdm>=4.62.0